fix: change output to debug level
